+++ backend/src/main/java/softeer/h9/hey/repository/Car/TagRepository.java
@@ -2,7 +2,6 @@ package softeer.h9.hey.repository.car;
 
 import java.util.List;
 
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
@@ -11,27 +10,12 @@ import org.springframework.stereotype.Repository;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
-import softeer.h9.hey.domain.archiving.SelectOptionTag;
 
 @Repository
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
 public class TagRepository {
 
 	private final NamedParameterJdbcTemplate jdbcTemplate;
-
-	public List<SelectOptionTag> findAllByArchivingIdSelectedOptions(final long id) {
-		String sql = "SELECT selectOption.id AS select_option_id, tag.content AS tag\n"
-			+ "FROM archiving\n"
-			+ "LEFT JOIN archiving_selectOption ON archiving.id = archiving_selectOption.archiving_id\n"
-			+ "LEFT JOIN selectOption ON archiving_selectOption.select_option_id = selectOption.id\n"
-			+ "LEFT JOIN selectOption_selectedTag ON archiving_selectOption.id = selectOption_selectedTag.archiving_selectOption_id\n"
-			+ "LEFT JOIN tag ON selectOption_selectedTag.tag_id = tag.id\n"
-			+ "WHERE archiving.id = :archivingId";
-		SqlParameterSource params = new MapSqlParameterSource()
-			.addValue("archivingId", id);
-
-		return jdbcTemplate.query(sql, params, selectOptionTagRowMapper());
-	}
 
 	public List<String> findTopByInteriorColorId(final int id, final int limit) {
 		String sql = "SELECT tag.content "
@@ -85,10 +69,6 @@ public class TagRepository {
 			.addValue("limit", limit);
 
 		return jdbcTemplate.query(sql, params, rowMapper());
-	}
-
-	private RowMapper<SelectOptionTag> selectOptionTagRowMapper() {
-		return BeanPropertyRowMapper.newInstance(SelectOptionTag.class);
 	}
 
 	private RowMapper<String> rowMapper() {
